🚚 move: SettingPage components -> sections
